requirements.txt
fastapi
uvicorn[standard]
python-multipart

sqlalchemy
psycopg2-binary
alembic

redis

passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]

pydantic
python-dotenv
shortuuid
qrcode
pillow

httpx

# testing

pytest
pytest-cov
pytest-asyncio
pytest-mock
pytest-env
httpx
locust